--- backend/src/modules/business/business.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { BusinessController } from './business.controller';
-
-describe('BusinessController', () => {
-  let controller: BusinessController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [BusinessController],
-    }).compile();
-
-    controller = module.get<BusinessController>(BusinessController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
